add compression for non generated images, add support for sized columns

+++ src/Service/MediaService.php
@@ -6,9 +6,11 @@ use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Form\FormInterface;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\Common\Collections\Criteria;
 use App\Entity\Media;
 use App\Entity\MediaAttachment;
 use App\Entity\Staff;
+use App\Enum\ImageSizeType;
 use Masterminds\HTML5;
 use Psr\Log\LoggerInterface;
 
@@ -22,14 +24,12 @@ class MediaService {
     private $mediumImageDimensions;
     private $largeImageDimensions;
     private $logger;
-
-    public const SMALL_IMAGE = 0;
-    public const MEDIUM_IMAGE = 1;
-    public const LARGE_IMAGE = 2;
+    private $uploadOriginalImage;
+    private $imageCompressionQuality;
 
     public function __construct(EntityManagerInterface $entityManager, FileNamerService $fileNamer, 
         LoggerInterface $logger, string $projectDir, string $uploadDestination, string $smallImageDimensions,
-        string $mediumImageDimensions, string $largeImageDimensions)
+        string $mediumImageDimensions, string $largeImageDimensions, bool $uploadOriginalImage, int $imageCompressionQuality)
     {
         $this->entityManager = $entityManager;
         $this->fileNamer = $fileNamer;
@@ -46,6 +46,9 @@ class MediaService {
 
         list($width, $height) = explode('x', $largeImageDimensions, 2);
         $this->largeImageDimensions = [(float)$width, (float)$height];;
+
+        $this->uploadOriginalImage = $uploadOriginalImage;
+        $this->imageCompressionQuality = $imageCompressionQuality;
     }
 
     /**
@@ -67,6 +70,7 @@ class MediaService {
             // move file to upload destination
             $name = $this->fileNamer->fileName($file);
             $subDirName = $this->fileNamer->directoryName($file);
+            $mimeType = $file->getMimeType();
             $publicDestination = join(DIRECTORY_SEPARATOR, [
                 $this->uploadDestination,
                 $subDirName
@@ -80,33 +84,72 @@ class MediaService {
             // do not upload until file name is unique
             /** @var \App\Repository\MediaRepository $mediaRepo */
             $mediaRepo = $this->entityManager->getRepository(Media::class);
-
+            
             while ($mediaRepo->findOneBy(['fileName' => $name]) !== null) {
                 $name = $this->fileNamer->fileName($file);
             }
 
-            // full absolute path of new file upload
-            $path = join(DIRECTORY_SEPARATOR, [
-                $absDestination,
-                $name
-            ]);
-
-            // move file to absolute upload destination from tmp directory
+            // move to upload destination and rename
             $file = $file->move($absDestination, $name);
+            
+            if (strpos($mimeType, "image/") !== false) {
+                 // Generate and upload the sized images
+                $sizedImages = $this->generateSizedImages($file);
+                $largeFile = null;
+                if (isset($sizedImages['large'])) {
+                    $largeFile = new File($sizedImages['large']);
+                    $largeFile = $largeFile->move($absDestination, $largeFile->getFilename());
+                }
 
-            return ['file' => $file, 'fileName' => $name, 'path' => $path, 'url' => $publicDestination . $name];
+                $mediumFile = null;
+                if (isset($sizedImages['medium'])) {
+                    $mediumFile = new File($sizedImages['medium']);
+                    $mediumFile = $mediumFile->move($absDestination, $mediumFile->getFilename());
+                }
+
+                $smallFile = null;
+                if (isset($sizedImages['small'])) {
+                    $smallFile = new File($sizedImages['small']);
+                    $smallFile = $smallFile->move($absDestination, $smallFile->getFilename());
+                }
+
+                $this->logger->info($sizedImages['small']);
+
+                if ($this->uploadOriginalImage === false) {
+                    // delete original image
+                    unlink($file->getRealPath());
+                    // reset reference to largest image version available
+                    $file = $largeFile ?? $mediumFile ?? $smallFile;
+                }
+            }
+
+            return [
+                'file' => $file,
+                'largeFile' => $largeFile,
+                'mediumFile' => $mediumFile,
+                'smallFile' => $smallFile,
+            ];
         } catch (\Exception $e) {
             // rollback the file upload; delete file 
-            if (isset($path) && file_exists($path)) {
-                unlink($path);
+            if (isset($file) && file_exists($file->getRealPath())) {
+                unlink($file->getRealPath());
+            }
+            if (isset($smallImage) && file_exists($smallImage->getRealPath())) {
+                unlink($smallImage->getRealPath());
+            }
+            if (isset($mediumImage) && file_exists($mediumImage->getRealPath())) {
+                unlink($mediumImage->getRealPath());
+            }
+            if (isset($largeImage) && file_exists($largeImage->getRealPath())) {
+                unlink($largeImage->getRealPath());
             }
             throw $e;
         }
     }
 
     private function generateSizedImage(File $file, int $sizeType) {
-        if ($sizeType !== self::SMALL_IMAGE && $sizeType !== self::MEDIUM_IMAGE
-                && $sizeType !== self::LARGE_IMAGE) {
+        if ($sizeType !== ImageSizeType::SMALL_IMAGE && $sizeType !== ImageSizeType::MEDIUM_IMAGE &&
+            $sizeType !== ImageSizeType::LARGE_IMAGE && $sizeType !== ImageSizeType::ORIGINAL_IMAGE) {
             throw new \Exception("Incorrect argument supplied for integer \$sizeType.");
         }
 
@@ -114,18 +157,18 @@ class MediaService {
         $fileName = pathinfo($file->getFilename(), PATHINFO_FILENAME);
         $fileExtension = $file->getExtension();
         
-        if ($sizeType == self::SMALL_IMAGE) {
-            $newWidth = $this->smallImageDimensions[0];
-            $newHeight = $this->smallImageDimensions[1];
-            $type = 'small';
-        } else if ($sizeType == self::MEDIUM_IMAGE) {
+        if ($sizeType === ImageSizeType::LARGE_IMAGE) {
+            $newWidth = $this->largeImageDimensions[0];
+            $newHeight = $this->largeImageDimensions[1];
+            $type = 'large';
+        } else if ($sizeType === ImageSizeType::MEDIUM_IMAGE) {
             $newWidth = $this->mediumImageDimensions[0];
             $newHeight = $this->mediumImageDimensions[1];
             $type = 'medium';
         } else {
-            $newWidth = $this->largeImageDimensions[0];
-            $newHeight = $this->largeImageDimensions[1];
-            $type = 'large';
+            $newWidth = $this->smallImageDimensions[0];
+            $newHeight = $this->smallImageDimensions[1];
+            $type = 'small';
         }
 
         $newFileName = sprintf('%s_%s.%s', $fileName, $type, $fileExtension);
@@ -134,12 +177,15 @@ class MediaService {
         if (copy($file->getRealPath(), $path) === true) {
             // Generate image
             $image = new \Imagick($path);
-            $image->setImageCompressionQuality(85);
+            $image->setImageCompressionQuality($this->imageCompressionQuality);
             $image->stripImage();
-            $res = $image->resizeImage($newWidth, 0, \Imagick::FILTER_LANCZOS, 1);
-            //$res = $image->thumbnailImage($newWidth, $newHeight, true);
-            //$res = $image->scaleImage($newWidth, 0);
-            
+            if ($sizeType === ImageSizeType::ORIGINAL_IMAGE) {
+                // only compress, keep original size
+                $res = $image->resizeImage($image->getImageWidth(), $image->getImageHeight(), \Imagick::FILTER_LANCZOS, 1);
+            } else {
+                $res = $image->resizeImage($newWidth, 0, \Imagick::FILTER_LANCZOS, 1);
+            }
+
             if ($res === true) {
                 $image->writeImage($path);
                 $image->destroy();
@@ -160,7 +206,11 @@ class MediaService {
                 return;
             }
 
-            $generatedImages = [];
+            $generatedImages = [
+                'large' => null,
+                'medium' => null,
+                'small' => null,
+            ];
 
             // Load the image file in Imagick
             $path = $file->getRealPath();
@@ -175,25 +225,25 @@ class MediaService {
             $sized = false;
             
             if ($width > $this->smallImageDimensions[0] || $height > $this->smallImageDimensions[1]) {
-                $generatedImages['small'] = $this->generateSizedImage($file, self::SMALL_IMAGE);
+                $generatedImages['small'] = $this->generateSizedImage($file, ImageSizeType::SMALL_IMAGE);
                 $sized = true;
             }
 
             if ($width > $this->mediumImageDimensions[0] || $height > $this->mediumImageDimensions[1]) {
-                $generatedImages['medium'] = $this->generateSizedImage($file, self::MEDIUM_IMAGE);
+                $generatedImages['medium'] = $this->generateSizedImage($file, ImageSizeType::MEDIUM_IMAGE);
                 $sized = true;
             }
 
             if ($width > $this->largeImageDimensions[0] || $height > $this->largeImageDimensions[1]) {
-                $generateImages['large'] = $this->generateSizedImage($file, self::LARGE_IMAGE);
+                $generateImages['large'] = $this->generateSizedImage($file, ImageSizeType::LARGE_IMAGE);
                 $sized = true;
             }
 
-            // if ($sized === false) {
-            //     // we did not generate any sized images
-            //     // perform only image compression
-            //     $image = new \Imagick($path);
-            // }
+            if ($sized === false) {
+                // we did not generate any sized images
+                // keep same image size, only compress
+                $generatedImages['small'] = $this->generateSizedImage($file, ImageSizeType::ORIGINAL_IMAGE);
+            }
             
             return $generatedImages;
         } catch (\Exception $e) {
@@ -204,7 +254,7 @@ class MediaService {
         }
     }
 
-    public function getRelativeUrlFromMedia(Media $media) { 
+    public function getRelativeUrlFromMedia(Media $media, int $sizeType = ImageSizeType::ORIGINAL_IMAGE) { 
         $mimeType = $media->getMimeType();
 
         if ($mimeType === null) return null;
@@ -214,11 +264,19 @@ class MediaService {
             $this->uploadDestination, 
             $subDirName
         ]);
-        $fileName = $media->getFileName();
+        if ($sizeType === ImageSizeType::LARGE_IMAGE) {
+            $fileName = $media->getLargeFileName();
+        } else if ($sizeType === ImageSizeType::MEDIUM_IMAGE) {
+            $fileName = $media->getMediumFileName();
+        } else if ($sizeType === ImageSizeType::SMALL_IMAGE) {
+            $fileName = $media->getSmallFileName();
+        } else {
+            $fileName = $media->getFileName();
+        }
+        
         $relativeUrl = sprintf("%s/%s", $publicDestination, $fileName);
 
         return $relativeUrl;
-
     }
 
     /**
@@ -303,9 +361,10 @@ class MediaService {
                 if (file_exists($filePath) !== false) {
                     $fileName = basename($filePath);
                     
+                    /** @var \App\Repository\MediaRepository $mediaRepo */
                     $mediaRepo = $this->entityManager->getRepository(Media::class);
                     /** @var Media $media */
-                    $media = $mediaRepo->findOneBy(['fileName' => $fileName]);
+                    $media = $mediaRepo->findOneByFileName($fileName);
                     
                     if ($media !== null) {
                         // check if media attachment already exists
@@ -404,7 +463,13 @@ class MediaService {
         foreach ($mediaAttachments as $attachment) {
             $media = $attachment->getMedia();
             $fileName = $media->getFileName();
-            if (isset($fileNames[$fileName]) === false) {
+            $largeFileName = $media->getLargeFileName() ?? $fileName;
+            $mediumFileName = $media->getMediumFileName() ?? $fileName;
+            $smallFileName = $media->getSmallFileName() ?? $fileName;
+            if (isset($fileNames[$fileName]) === false &&
+                isset($fileNames[$largeFileName]) === false &&
+                isset($fileNames[$mediumFileName]) === false &&
+                isset($fileNames[$smallFileName]) === false) {
                 $attachment->setMedia(null);
                 $this->entityManager->remove($attachment);
             }
